refactor(JsonSchemaEditor): 重构高级设置和导入功能
-将高级设置和导入功能抽离为独立的组件
- 优化高级设置的表单结构和逻辑
- 改进导入功能,支持 JSON 和 JSON Schema 两种格式
- 调整 SchemaItem 组件,集成新的高级设置和导入组件
- 修复重命名属性时 required 字段未同步更新的问题
