+++ src/JsonSchemaEditor/SchemaItem/index.tsx
@@ -2,16 +2,11 @@ import {
   Button,
   Checkbox,
   Dropdown,
-  Form,
   Grid,
   Input,
-  InputNumber,
   Menu,
   Message,
-  Modal,
-  Radio,
   Select,
-  Switch,
   Tooltip,
 } from '@arco-design/web-react';
 import {
@@ -22,20 +17,15 @@ import {
   IconSettings,
   IconUpload,
 } from '@arco-design/web-react/icon';
-import { Draft07 } from 'json-schema-library';
-import _ from 'lodash';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import MonacoEditor from '../MonacoEditor';
+import ImportModal from '@quiet-front-end/json-schema-editor-arco/JsonSchemaEditor/SchemaItem/ImportModal';
+import React, { useEffect, useState } from 'react';
 import { JSONSchema7 } from '../types';
 import {
   SchemaTypeOptions,
-  SchemaTypes,
-  StringFormat,
   getDefaultSchema,
   getPropertyIndex,
-  parseJsonStr,
-  resolveJsonSchemaRef,
 } from '../utils';
+import AdvancedSettingModal from './AdvancedSettingModal';
 
 const Row = Grid.Row;
 const Col = Grid.Col;
@@ -70,7 +60,6 @@ type SchemaItemProps = {
 };
 
 function SchemaItem(props: SchemaItemProps) {
-  const [advancedForm] = Form.useForm();
   const {
     changeSchema,
     renameProperty,
@@ -84,7 +73,6 @@ function SchemaItem(props: SchemaItemProps) {
   } = props;
 
   const [schema, setSchema] = useState(props.schema);
-  const [formSchema, setFormSchema] = useState<any>();
   const [propertyName, setPropertyName] = useState(props.propertyName);
   const [schemaTitle, setSchemaTitle] = useState(schema.title);
   const [schemaDescription, setSchemaDescription] = useState(
@@ -99,15 +87,6 @@ function SchemaItem(props: SchemaItemProps) {
   const [expand, setExpand] = useState(true);
   const [advancedModal, setAdvancedModal] = useState(false);
   const [importModalVisible, setImportModalVisible] = useState(false);
-  const [importType, setImportType] = useState<'json' | 'json-schema'>('json');
-  const [importValue, setImportValue] = useState<string | undefined>();
-  const [isObject, setIsObject] = useState(false);
-  const [isArray, setIsArray] = useState(false);
-  const [isNumber, setIsNumber] = useState(false);
-  const [isBoolean, setIsBoolean] = useState(false);
-  const [isInteger, setIsInteger] = useState(false);
-  const [isString, setIsString] = useState(false);
-  const editorRef = useRef<any>(null);
   const isRoot = typeof propertyName === 'undefined';
 
   useEffect(() => {
@@ -122,40 +101,6 @@ function SchemaItem(props: SchemaItemProps) {
     setNodeDepth(props.nodeDepth ? props.nodeDepth : 0);
   }, [props.nodeDepth]);
 
-  const handleDebounce = useCallback(
-    _.debounce(
-      (callback) => {
-        if (typeof callback === 'function') {
-          callback();
-        } else {
-          console.log('Provided argument is not a function');
-        }
-      },
-      300,
-      { maxWait: 1000 },
-    ),
-    [],
-  );
-
-  useEffect(() => {
-    return () => {
-      handleDebounce.cancel();
-    };
-  }, [handleDebounce]);
-
-  useEffect(() => {
-    if (!advancedModal || !formSchema) {
-      return;
-    }
-    advancedForm.setFieldsValue(formSchema);
-    setIsObject(formSchema.type === 'object');
-    setIsArray(formSchema.type === 'array');
-    setIsNumber(formSchema.type === 'number');
-    setIsBoolean(formSchema.type === 'boolean');
-    setIsInteger(formSchema.type === 'integer');
-    setIsString(formSchema.type === 'string');
-  }, [advancedModal, JSON.stringify(formSchema)]);
-
   const schemaItems: any = schema.items;
 
   const addChildNode: boolean =
@@ -165,10 +110,6 @@ function SchemaItem(props: SchemaItemProps) {
     ) &&
     !isArrayItems &&
     !isRoot;
-
-  function handleEditorDidMount(editor: any) {
-    editorRef.current = editor;
-  }
 
   if (!schema.type) {
     return <></>;
@@ -301,8 +242,7 @@ function SchemaItem(props: SchemaItemProps) {
                   ) {
                     return;
                   }
-                  setFormSchema(schema);
-                  setAdvancedModal(!advancedModal);
+                  setAdvancedModal(true);
                 }}
               />
             </Tooltip>
@@ -354,7 +294,7 @@ function SchemaItem(props: SchemaItemProps) {
                 </Tooltip>
               </Dropdown>
             ) : (
-              <div style={{ width: '24px' }} />
+              <div />
             )}
             <Col flex={'24px'}>
               {isRoot ? (
@@ -382,13 +322,11 @@ function SchemaItem(props: SchemaItemProps) {
                   />
                 </Tooltip>
               ) : (
-                <div style={{ width: '24px' }} />
+                <div />
               )}
             </Col>
             {isRoot && schema.type !== 'object' && (
-              <Col flex={'24px'}>
-                {!isArrayItems && <div style={{ width: '24px' }} />}
-              </Col>
+              <Col flex={'24px'}>{!isArrayItems && <div />}</Col>
             )}
           </Row>
         </Col>
@@ -432,459 +370,34 @@ function SchemaItem(props: SchemaItemProps) {
           handleAdvancedSettingClick={handleAdvancedSettingClick}
         />
       )}
-      <Modal
-        title="高级设置"
-        style={{ width: 900 }}
+      <AdvancedSettingModal
+        schema={schema}
         visible={advancedModal}
-        okText={'保存'}
-        cancelText={'取消'}
-        onOk={async () => {
+        onOk={async (newSchema) => {
           if (!changeSchema) {
             return;
           }
           if (isRoot || schema.type === 'object') {
-            changeSchema(namePath, { ...schema, ...formSchema });
-            setAdvancedModal(!advancedModal);
+            changeSchema(namePath, { ...newSchema });
+            setAdvancedModal(false);
             return;
           }
-          try {
-            await advancedForm.validate();
-            changeSchema(
-              namePath,
-              { ...schema, ...formSchema, ...advancedForm.getFieldsValue() },
-              propertyName,
-            );
-            setAdvancedModal(!advancedModal);
-          } catch (e) {
-            console.log(advancedForm.getFieldsError());
-            Message.error('字段校验失败，请检查字段！');
-          }
+          changeSchema(namePath, { ...newSchema }, propertyName);
+          setAdvancedModal(false);
         }}
-        onCancel={() => setAdvancedModal(!advancedModal)}
-      >
-        <Form
-          form={advancedForm}
-          onValuesChange={(_, allValues) => {
-            if (editorRef.current) {
-              editorRef.current.setValue(
-                JSON.stringify({ ...formSchema, ...allValues }, null, 2),
-              );
-            }
-          }}
-        >
-          {!isObject && SchemaTypes.indexOf(formSchema?.type) !== -1 && (
-            <div
-              style={{
-                borderLeft: `3px solid rgb(var(--primary-6))`,
-                fontSize: 16,
-                fontWeight: 399,
-                paddingLeft: 8,
-                marginBottom: 13,
-              }}
-            >
-              基本设置
-            </div>
-          )}
-          {(isString || isNumber || isInteger || isBoolean) && (
-            <Row
-              justify={'start'}
-              align={'center'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                默认值：
-              </Col>
-              <Col span={8}>
-                {isString && (
-                  <Form.Item noStyle field={'default'}>
-                    <Input
-                      style={{ width: '100%' }}
-                      placeholder={'请输入默认值'}
-                    />
-                  </Form.Item>
-                )}
-                {(isNumber || isInteger) && (
-                  <Form.Item noStyle field={'default'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入默认值'}
-                    />
-                  </Form.Item>
-                )}
-                {isBoolean && (
-                  <Form.Item noStyle field={'default'}>
-                    <Select
-                      style={{ width: '100%' }}
-                      placeholder={'请选择默认值'}
-                      options={[
-                        { value: 'true', label: 'true' },
-                        { value: 'false', label: 'false' },
-                      ]}
-                    />
-                  </Form.Item>
-                )}
-              </Col>
-            </Row>
-          )}
-          {isString && (
-            <Row
-              justify={'start'}
-              align={'center'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                最小长度：
-              </Col>
-              <Col span={8}>
-                <Form.Item noStyle field={'minLength'}>
-                  <InputNumber
-                    min={0}
-                    style={{ width: '100%' }}
-                    parser={(value) =>
-                      value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                    }
-                    formatter={(value) =>
-                      value ? `${Math.floor(Math.max(Number(value), 0))}` : ''
-                    }
-                    placeholder={'请输入最小长度'}
-                  />
-                </Form.Item>
-              </Col>
-              <Col span={4} style={{ textAlign: 'right' }}>
-                最大长度：
-              </Col>
-              <Col span={8}>
-                <Form.Item noStyle field={'maxLength'}>
-                  <InputNumber
-                    min={0}
-                    style={{ width: '100%' }}
-                    parser={(value) =>
-                      value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                    }
-                    formatter={(value) =>
-                      value ? `${Math.floor(Math.max(Number(value), 0))}` : ''
-                    }
-                    placeholder={'请输入最大长度'}
-                  />
-                </Form.Item>
-              </Col>
-            </Row>
-          )}
-          {(isNumber || isInteger) && (
-            <>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最小值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'minimum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入最小值'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最大值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'maximum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入最大值'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  排他最小值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'exclusiveMinimum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入排他最小值'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  排他最大值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'exclusiveMaximum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入排他最大值'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {isString && (
-            <>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  正则匹配：
-                </Col>
-                <Col span={20}>
-                  <Form.Item noStyle field={'pattern'}>
-                    <Input placeholder={'请输入正则匹配公式'} />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  格式：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'format'}>
-                    <Select
-                      allowClear
-                      options={StringFormat.map((sf) => sf.value)}
-                      placeholder={'请选择字符串格式'}
-                      style={{ width: '100%' }}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {isArray && (
-            <>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  元素唯一：
-                </Col>
-                <Col span={20}>
-                  <Form.Item
-                    noStyle
-                    field={'uniqueItems'}
-                    triggerPropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'center'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最少元素个数：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'minItems'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      parser={(value) =>
-                        value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                      }
-                      formatter={(value) =>
-                        value ? `${Math.floor(Math.max(Number(value), 0))}` : ''
-                      }
-                      placeholder={'请输入最少元素个数'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最多元素个数：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle field={'maxItems'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      parser={(value) =>
-                        value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                      }
-                      formatter={(value) =>
-                        value ? `${Math.floor(Math.max(Number(value), 0))}` : ''
-                      }
-                      placeholder={'请输入最多元素个数'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {(isString || isNumber || isInteger) && (
-            <Row
-              justify={'start'}
-              align={'center'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                枚举：
-              </Col>
-              <Col span={20}>
-                <Form.List field="enums">
-                  {(fields, { add, remove }) => (
-                    <>
-                      <Row gutter={10}>
-                        {fields.map(({ key, field }, index) => {
-                          return (
-                            <Col span={12} key={key}>
-                              <Row
-                                justify={'start'}
-                                align={'center'}
-                                style={{ marginBottom: 6 }}
-                              >
-                                <Col flex={'auto'}>
-                                  {isString && (
-                                    <Form.Item
-                                      noStyle
-                                      field={field}
-                                      rules={[{ required: true }]}
-                                    >
-                                      <Input placeholder="请输入枚举值" />
-                                    </Form.Item>
-                                  )}
-                                  {(isNumber || isInteger) && (
-                                    <Form.Item
-                                      noStyle
-                                      field={field}
-                                      rules={[{ required: true }]}
-                                    >
-                                      <InputNumber
-                                        style={{ width: '100%' }}
-                                        placeholder="请输入枚举值"
-                                      />
-                                    </Form.Item>
-                                  )}
-                                </Col>
-                                <Col flex={'36px'} style={{ paddingLeft: 7 }}>
-                                  <Button
-                                    icon={<IconDelete />}
-                                    shape="circle"
-                                    status="danger"
-                                    onClick={() => remove(index)}
-                                  />
-                                </Col>
-                              </Row>
-                            </Col>
-                          );
-                        })}
-                      </Row>
-                      <Row>
-                        <Col span={12}>
-                          <Form.Item noStyle>
-                            <Button onClick={() => add()} icon={<IconPlus />}>
-                              添加枚举值
-                            </Button>
-                          </Form.Item>
-                        </Col>
-                      </Row>
-                    </>
-                  )}
-                </Form.List>
-              </Col>
-            </Row>
-          )}
-          <div
-            style={{
-              borderLeft: `3px solid rgb(var(--primary-6))`,
-              fontSize: 16,
-              fontWeight: 399,
-              paddingLeft: 8,
-              marginBottom: 13,
-            }}
-          >
-            Json Schema
-          </div>
-          <MonacoEditor
-            height={300}
-            language="json"
-            value={JSON.stringify(formSchema, null, 2)}
-            handleEditorDidMount={handleEditorDidMount}
-            onChange={(value) => {
-              handleDebounce(() => {
-                if (value) {
-                  try {
-                    const editorSchema = JSON.parse(value);
-                    setFormSchema(editorSchema);
-                  } catch (e) {}
-                }
-              });
-            }}
-          />
-        </Form>
-      </Modal>
+        onCancel={() => setAdvancedModal(false)}
+      />
 
-      <Modal
-        title="导入"
-        style={{ width: 900 }}
-        okText={'导入'}
-        cancelText={'取消'}
-        visible={importModalVisible}
-        onOk={async () => {
-          if (!importValue || importValue.length === 0) {
-            Message.warning('请输入导入的 Json 数据');
-            return;
-          }
-          const importObject = parseJsonStr(importValue);
-          if (!importObject) {
-            Message.error('导入的内容不是 Json 格式的数据');
-            return;
-          }
-          let schema;
-          switch (importType) {
-            case 'json':
-              schema = new Draft07().createSchemaOf(importObject);
-              break;
-            case 'json-schema':
-              schema = await resolveJsonSchemaRef(importObject);
-              break;
-          }
+      <ImportModal
+        open={importModalVisible}
+        onOk={(newSchema) => {
           if (changeSchema) {
-            changeSchema([], schema);
-            setImportModalVisible(!importModalVisible);
-            setImportValue(undefined);
+            changeSchema([], newSchema);
+            setImportModalVisible(false);
           }
         }}
-        onCancel={() => setImportModalVisible(!importModalVisible)}
-      >
-        <Row style={{ marginBottom: 16 }}>
-          <Radio.Group
-            value={importType}
-            type="button"
-            onChange={(type) => setImportType(type)}
-            options={[
-              { value: 'json', label: 'Json' },
-              { value: 'json-schema', label: 'JsonSchema' },
-            ]}
-          />
-        </Row>
-        <Row>
-          <MonacoEditor
-            height={399}
-            language="json"
-            handleEditorDidMount={handleEditorDidMount}
-            onChange={(value) => setImportValue(value)}
-          />
-        </Row>
-      </Modal>
+        onCancel={() => setImportModalVisible(false)}
+      />
     </>
   );
 }
